+++ src/main/scala/io/viash/config/arguments/FileArgument.scala
@@ -47,6 +47,18 @@ case class FileArgument(
   @default("Empty")
   alternatives: OneOrMore[String] = Nil,
 
+  @description("A clean notation of the argument's name. This is used when generating documentation.")
+  @example("label: an argument", "yaml")
+  @default("Empty")
+  @since("Viash 0.9.0")
+  label: Option[String] = None,
+
+  @description("A short summary of the argument. This is used when generating documentation.")
+  @example("summary: Use this argument to do X", "yaml")
+  @default("Empty")
+  @since("Viash 0.9.0")
+  summary: Option[String] = None,
+
   @description("A description of the argument. This will be displayed with `--help`.")
   description: Option[String] = None,
 
@@ -165,6 +177,8 @@ case class FileArgument(
     `type`: String, 
     name: String, 
     alternatives: OneOrMore[String],
+    label: Option[String],
+    summary: Option[String],
     description: Option[String],
     info: Json,
     example: OneOrMore[Path],
@@ -176,7 +190,7 @@ case class FileArgument(
     dest: String
   ): Argument[Path] = {
     copy(
-      name, alternatives, description, info, example, default, 
+      name, alternatives, label, summary, description, info, example, default, 
       this.must_exist, this.create_parent, required, direction, 
       multiple, multiple_sep, dest, `type`
     )