+++ src/main/scala/io/viash/config/arguments/LongArgument.scala
@@ -47,6 +47,18 @@ case class LongArgument(
   @default("Empty")
   alternatives: OneOrMore[String] = Nil,
 
+  @description("A clean notation of the argument's name. This is used when generating documentation.")
+  @example("label: an argument", "yaml")
+  @default("Empty")
+  @since("Viash 0.9.0")
+  label: Option[String] = None,
+
+  @description("A short summary of the argument. This is used when generating documentation.")
+  @example("summary: Use this argument to do X", "yaml")
+  @default("Empty")
+  @since("Viash 0.9.0")
+  summary: Option[String] = None,
+
   @description("A description of the argument. This will be displayed with `--help`.")
   description: Option[String] = None,
 
@@ -150,6 +162,8 @@ case class LongArgument(
     `type`: String, 
     name: String, 
     alternatives: OneOrMore[String],
+    label: Option[String],
+    summary: Option[String],
     description: Option[String],
     info: Json,
     example: OneOrMore[Long],
@@ -160,6 +174,6 @@ case class LongArgument(
     multiple_sep: String,
     dest: String
   ): Argument[Long] = {
-    copy(name, alternatives, description, info, example, default, required, this.choices, this.min, this.max, direction, multiple, multiple_sep, dest, `type`)
+    copy(name, alternatives, label, summary, description, info, example, default, required, this.choices, this.min, this.max, direction, multiple, multiple_sep, dest, `type`)
   }
 }